Time: 0 ms (100%), Space: 9 MB (44.12%) - LeetHub

+++ 1156-occurrences-after-bigram/1156-occurrences-after-bigram.cpp
@@ -9,7 +9,7 @@ public:
            vec.push_back(a);
         }
         vector<string>ans;
-        for (int i=0;i<vec.size()-2;i++){
+        for (size_t i=0;i<vec.size()-2;i++){
             if(vec[i]==first && vec[i+1]==second){
                      ans.push_back(vec[i+2]);
             }
